Reorganized theme file.
